Print Position Element Asterisk (Histogram)
This algorithm receive ten values to will stored within of each position of array. So, the algoritm print the position, the element within of posistion and amount of asterisk contained of position. For exemple, if the algorithm to receive the value five in the first position, the algorithm will print: position 0, element 5, and five asterisk. 
The idea this algorithm is to do a Histogram of asterisk.