Receive two String and print
This algorithm receive two string, a by method "scanf" and other by method "gets", and print the two words. The idea this algorithm is receive the name by method "scanf", and receive the last name by method "gets". So, the algorithm print the name, and the last name on the same line.